requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langgraph>=0.1.0
ragas>=0.1.9
openai>=1.30.0
faiss-cpu>=1.7.4
tiktoken>=0.7.0
python-dotenv>=1.0.0
datasets>=2.19.0
pandas>=2.0.0
